src/main/java/com/example/order/utils/MapProductToJsonConverter.java
package com.example.order.utils;

import com.example.order.entity.Product;
import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.core.type.TypeReference;
import com.fasterxml.jackson.databind.ObjectMapper;
import jakarta.persistence.AttributeConverter;
import jakarta.persistence.Converter;

import java.io.IOException;
import java.util.Map;

@Converter
public class MapProductToJsonConverter implements AttributeConverter<Map<String, Product>, String> {

    private final ObjectMapper objectMapper = new ObjectMapper();

    @Override
    public String convertToDatabaseColumn(Map<String, Product> attribute) {
        try {
            return objectMapper.writeValueAsString(attribute);
        } catch (JsonProcessingException e) {
            throw new IllegalArgumentException("Error converting product map to JSON", e);
        }
    }

    @Override
    public Map<String, Product> convertToEntityAttribute(String dbData) {
        try {
            return objectMapper.readValue(dbData, new TypeReference<Map<String, Product>>() {});
        } catch (IOException e) {
            throw new IllegalArgumentException("Error converting product JSON to map", e);
        }
    }
}
